requirements.txt
numpy~=1.21.1
requests~=2.31.0
pyfacer~=0.0.4
timm~=0.5.4
aiofiles~=23.2.1
aiosqlite~=0.19.0
uvicorn~=0.26.0
fastapi~=0.109.0
python-dotenv~=1.0.0
torch~=2.1.2
pillow~=10.2.0
python-telegram-bot~=20.7
torchvision~=0.16.2
tqdm~=4.66.1
scikit-image~=0.22.0
insightface~=0.2.1
imageio~=2.4.1
huggingface-hub~=0.20.1
validators~=0.22.0
icecream~=2.1.3